requirements.txt
asgiref==3.3.4
astroid==2.5.3
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.4
Django==3.2
django-livereload-server==0.3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Flask==1.1.1
idna==2.8
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.2
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.0.1
pylint==2.7.4
pytz==2021.1
requests==2.22.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.12.1
